requirements.txt
Flask
Flask-Cors
python-dotenv
googlemaps
gtts
langchain
langchain-community
langchain-google-genai
langchain-chroma
google-generativeai
chromadb
